+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [UNRELEASED]
+### Fixed
+- #1410 error with repeat alarm config field while not exists after update
+
 ## [1.1.0] - 2023-03-13
 ### Added
 - Send unscheduled alarm event